chore: move core::string_utils to core::string::utils
Also move case tables to a separate file.
