+++ core/src/lib.rs
@@ -43,7 +43,7 @@ pub mod matrix;
 mod player;
 mod prelude;
 pub mod shape_utils;
-pub mod string_utils;
+pub mod string;
 pub mod tag_utils;
 mod transform;
 mod types;
