+++ core/src/string.rs
@@ -1,0 +1,2 @@
+mod tables;
+pub mod utils;
